watchmate/requirements.txt.txt
asgiref==3.8.1
Django==5.2.3
django-filter==25.1
django_db==0.0.7
djangorestframework==3.16.0
Pygments==2.19.2
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.2